refactor: Move downloading & notifications out of background script
